Fix bwc issues with materialized CTEs (#46)
This cherry picks the functionality from duckdb/duckdb#19420

Follow-up from duckdb/duckdb#19393

There are a number of issues still caused by serializing CTE nodes - this PR makes it so that we only serialize CTE nodes when MATERIALIZED is explicitly defined, and serialize only the CommonTableExpressionMap otherwise. In addition, we never deserialize CTENodes anymore - and always reconstruct them from the CommonTableExpressionMap.
---------

Co-authored-by: Mark <mark.raasveldt@gmail.com>